+++ docs/source/changelog.rst
@@ -2,6 +2,11 @@
 Change Log
 **********
 
+**Pyro 4.43**
+
+- improved docs on instance modes and instance creation
+
+
 **Pyro 4.42**
 
 - added dill serialization support (https://pypi.python.org/pypi/dill)